+++ src/runtime/types/prose.ts
@@ -1,6 +1,6 @@
 export * from '../components/prose/A.vue'
-// export * from '../components/prose/Accordion.vue'
-// export * from '../components/prose/AccordionItem.vue'
+export * from '../components/prose/Accordion.vue'
+export * from '../components/prose/AccordionItem.vue'
 // export * from '../components/prose/Badge.vue'
 export * from '../components/prose/Blockquote.vue'
 export * from '../components/prose/Callout.vue'