+++ src/theme/prose/accordion.ts
@@ -1,0 +1,11 @@
+/**
+ * ProseAccordion
+ * ---
+ */
+
+export default {
+  slots: {
+    root: 'my-[20px]',
+    trigger: 'text-(--b24ui-typography-label-color)'
+  }
+}